% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crabrawlR_fun.R
\name{getcounts}
\alias{getcounts}
\title{Get counts how many ortholog groups are in each COG category}
\usage{
getcounts(df_sub, categories, thaumtab)
}
\arguments{
\item{vector}{of (character) names of ortholog groups to get functional annotation counts for}

\item{vector}{of (character) names of COG categories}

\item{dataframe}{of COG annotations for pangenome}
}
\value{
vector containing # of OF in each category
}
\description{
Get counts how many ortholog groups are in each COG category
}
\examples{
getcounts(orthologname_list, cogcategories, annotation_df)
}
\keyword{COG}
\keyword{categories,}
\keyword{counts}
\keyword{enrichment,}
\keyword{functional}
