% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crabrawlR_fun.R
\name{getmissing}
\alias{getmissing}
\title{Interrogate which taxa are missing in a list of gene families}
\usage{
getmissing(pmatrix, taxlist_yes, genefamily_list)
}
\arguments{
\item{dataframe}{formatted by panpretty and NAcheck}

\item{vector}{of row numbers for list of taxa to include}

\item{vector}{of gene family names (product of sharedgroups or getunique)}
}
\value{
numeric int, max # of taxa that can have genes missing
}
\description{
Interrogate which taxa are missing in a list of gene families
}
\examples{
getmissing(pangenome_matrix, taxa_yes, genefamily_list)
}
\keyword{matrix,}
\keyword{pangenome,}
\keyword{shared,}
\keyword{specific,}
\keyword{unique}
