% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crabrawlR_fun.R
\name{getunique}
\alias{getunique}
\title{Find gene families present ONLY IN (unique to) a list of taxa}
\usage{
getunique(df, r_yes, r_no)
}
\arguments{
\item{dataframe}{formatted by panpretty and NAcheck}

\item{vector}{of row numbers for list of taxa to include}

\item{vector}{of row numbers for list of taxa to exclude}
}
\value{
numeric int, max # of taxa that can have genes missing
}
\description{
Find gene families present ONLY IN (unique to) a list of taxa
}
\examples{
getunique(pangenome_matrix, taxa_yes, taxa_no)
}
\keyword{matrix,}
\keyword{pangenome,}
\keyword{shared,}
\keyword{specific,}
\keyword{unique}
