% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crabrawlR_fun.R
\name{NAcheck}
\alias{NAcheck}
\title{Remove columns with NAs and return 'clean' matrix.
For get_homologues generated pangenome matrices.}
\usage{
NAcheck(df)
}
\arguments{
\item{dataframe}{generated by panpretty()}
}
\value{
clean dataframe with no NAs
}
\description{
Remove columns with NAs and return 'clean' matrix.
For get_homologues generated pangenome matrices.
}
\examples{
NAcheck(pangenomeMatrix)
}
\keyword{matrix,}
\keyword{pangenome,}
\keyword{removeNA}
