% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crabrawlR_fun.R
\name{panpretty}
\alias{panpretty}
\title{Set rownames as taxa names, colnames as genefam names.
For get_homologues generated pangenome matrices.}
\usage{
panpretty(df)
}
\arguments{
\item{dataframe}{of matrix imported using read.csv}
}
\value{
dataframe with rownames as taxa names, colnames as genefam names
}
\description{
Set rownames as taxa names, colnames as genefam names.
For get_homologues generated pangenome matrices.
}
\examples{
panpretty(pangenomeMatrix)
}
\keyword{matrix}
\keyword{pangenome,}
