% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crabrawlR_fun.R
\name{sharedgroups}
\alias{sharedgroups}
\title{Find shared gene families between a list of taxa}
\usage{
sharedgroups(df, r_yes)
}
\arguments{
\item{dataframe}{formatted by panpretty and NAcheck}

\item{vector}{of row numbers for list of taxa to include}
}
\value{
numeric int, max # of taxa that can have genes missing
}
\description{
Find shared gene families between a list of taxa
}
\examples{
sharedgroups(pangenomeMatrix_clean, tax_yes)
}
\keyword{matrix,}
\keyword{pangenome,}
\keyword{shared,}
\keyword{specific,}
\keyword{unique}
