% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crabrawlR_fun.R
\name{shellnum}
\alias{shellnum}
\title{Shell lower bound value from Van Vliet 2017 (15% of taxa)
Returned int is max \# of taxa that can have zeros in column (gene family) to qualify as 'shell'}
\usage{
shellnum(taxct)
}
\arguments{
\item{vector}{with taxa names or taxa row nums}
}
\value{
numeric int, max # of taxa that can have genes missing
}
\description{
Shell lower bound value from Van Vliet 2017 (15% of taxa)
Returned int is max \# of taxa that can have zeros in column (gene family) to qualify as 'shell'
}
\examples{
shellnum(PangenomeMatrix[,1])
}
\keyword{matrix}
\keyword{pangenome,}
