% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crabrawlR_fun.R
\name{softnum}
\alias{softnum}
\title{Soft core taxa value threshold--95\% from Kaas et al 2012
Up to what \% of taxa can have this gene absent to qualify for 'soft-core' definition?.
Returned int is \# of taxa that can have zeros in column (gene family) to qualify as 'soft-core'.}
\usage{
softnum(taxct)
}
\arguments{
\item{vector}{with taxa names or taxa row nums}
}
\value{
numeric int, max # of taxa that can have genes missing
}
\description{
Soft core taxa value threshold--95\% from Kaas et al 2012
Up to what \% of taxa can have this gene absent to qualify for 'soft-core' definition?.
Returned int is \# of taxa that can have zeros in column (gene family) to qualify as 'soft-core'.
}
\examples{
softnum(PangenomeMatrix[,1])
}
\keyword{matrix}
\keyword{pangenome,}
