% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crabrawlR_fun.R
\name{taxrownums}
\alias{taxrownums}
\title{Given a list of taxa names (that are also rownames), get rownumbers}
\usage{
taxrownums(df = df, taxa_yes = NULL, taxa_no = NULL)
}
\arguments{
\item{dataframe}{formatted by panpretty and NAcheck}

\item{vector}{of rownames for list of taxa to include (optional, default value NULL)}

\item{vector}{of rownames for list of taxa to exclude (optional, default value NULL)}
}
\value{
list with rownumbers for taxa to include ($r_yes) or exclude ($r_no)
}
\description{
Given a list of taxa names (that are also rownames), get rownumbers
}
\examples{
taxrownums(df=myPangenomeMatrix, taxa_yes=taxa_yes, taxa_no=taxa_no)
}
\keyword{rownames}
\keyword{rownums,}
